Replace own vector search with Chromem-go
Because my implementation is apparently not producing the expected
results and this project fits the bill quite nicely.

Most of the tests are failing, but the core functionality appears to
work.

- [x] Set
- [ ] Get
- [ ] Delete (we have to delete the whole collection presently)
- [ ] Query (it returns normalized embeddings, not the originals)
- [x] Query Normalized

Some features are missing from Chromem, but we don't strictly need
them. Meanwhile there are maybe some minor things that need adding on
our side to get the basic functionality working.